feat: v1: add v1 router and update main.py
